--- docs/investigations/storage/beego/beego.md
@@ -1,6 +1,0 @@
-# Beego Investigation
-
-> Beego ORM is a powerful ORM framework written in Go. It is inspired by Django ORM and SQLAlchemy.
-> This framework is still under development so compatibility is not guaranteed.
-
-https://beego.me/docs/mvc/model/overview.md
