--- docs/investigations/storage/gorm/gorm.md
@@ -1,3 +1,0 @@
-# GORM Investigation
-
-https://www.reddit.com/r/golang/comments/8j3219/anyone_using_gorm_in_production_is_it_slow/
