+++ docs/investigations/storage/sql-toolbox/libs/gorm/internal/model/model.go
