+++ docs/investigations/storage/sql-toolbox/libs/sqlx/internal/db/id.go
