+++ docs/investigations/storage/sql-toolbox/libs/sqlx/internal/domain/application/dao_test.go
@@ -1,0 +1,14 @@
+package application
+
+import (
+	"fmt"
+	sq "github.com/Masterminds/squirrel"
+	"github.com/kyma-incubator/compass/docs/investigations/storage/sql-toolbox/libs/sqlxtoolbox/libs/sqlx/dto"
+	"testing"
+)
+
+func TestAbd(t *testing.T) {
+	appDTO := dto.ApplicationDTO{}
+	err := sq.Insert("applications").Scan(appDTO).Error()
+	fmt.Println("err",err)
+}
