+++ docs/investigations/storage/sql-toolbox/libs/sqlx/internal/model/model.go
